pyproject.toml
[project]
name = "my-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.0",
    "chainlit>=2.9.4",
    "cos-python-sdk-v5>=1.9.41",
    "deepagents>=0.3.9",
    "docx2txt>=0.9",
    "dotenv>=0.9.9",
    "ebooklib>=0.20",
    "faiss-cpu>=1.13.2",
    "faker>=39.0.0",
    "fastapi>=0.128.0",
    "google-genai>=1.56.0",
    "gradio>=6.2.0",
    "ipython>=8.37.0",
    "langchain>=1.2.0",
    "langchain-deepseek>=1.0.1",
    "langchain-experimental>=0.4.1",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-milvus>=0.3.3",
    "langchain-neo4j>=0.8.0",
    "langchain-openai>=1.1.5",
    "langchain-tavily>=0.2.16",
    "langgraph>=1.0.5",
    "langgraph-checkpoint-sqlite>=3.0.1",
    "langgraph-cli[inmem]>=0.4.11",
    "libmagic>=1.0",
    "neo4j>=6.1.0",
    "pillow>=12.0.0",
    "pip>=25.3",
    "playwright>=1.57.0",
    "pymysql>=1.1.2",
    "pypdf2>=3.0.1",
    "python-magic>=0.4.27",
    "rank-bm25>=0.2.2",
    "rich>=14.2.0",
    "selenium>=4.40.0",
    "sqlalchemy>=2.0.45",
    "streamlit>=1.52.2",
    "tencentcloud-sdk-python==3.0.1120",
    "unstructured[local-inference]>=0.18.24",
    "uv>=0.9.18",
    "uvicorn>=0.40.0",
    "webdriver-manager>=4.0.2",
]
